Extract TurnEngine service object from Game god object (#6)
All turn execution and query logic extracted from Game into a new plain
Ruby service class at app/services/turn_engine.rb. Game retains only AR
model responsibilities: associations, lifecycle (start/waiting?/playing?),
board instantiation, broadcasting, and snapshot/replay.

Controller and views updated to instantiate TurnEngine.new(@game) directly
(no delegation stubs). The engine: local is threaded through the view partial
chain for the query methods (turn_state, turn_endable?, undo_allowed?,
tile_activatable?) used in rendering. Five new integration tests added in
test/services/turn_engine_test.rb; duplicated turn tests removed from
game_lifecycle_test.rb.

Game model reduced from ~617 to ~185 lines.
