+++ app/views/games/_players.html.erb
@@ -26,7 +26,7 @@
     <% end %>
     <div class="player-area">
       <%= content_tag :div, id: "game_player_#{player.id}" do %>
-        <%= render partial: "games/game_player", locals: { game:, player:, n: } %>
+        <%= render partial: "games/game_player", locals: { game:, player:, n:, engine: } %>
       <% end %>
     </div>
   <% end %>
