+++ app/views/games/_turn_state.html.erb
@@ -1,4 +1,4 @@
-<%= game.turn_state %>
+<%= engine.turn_state %>
 <span class="mandatory-count"><%= game.mandatory_count %></span>
 <span id="current-action"
   data-type="<%= game.current_action&.dig('type') %>"
