+++ test/models/game_test.rb
@@ -32,7 +32,7 @@ class GameTest < ActiveSupport::TestCase
     game.save
 
     # Simulate end of turn
-    game.end_turn
+    engine(game).end_turn
 
     # Check that the deck is shuffled and discard is cleared
     assert_equal [], game.discard
@@ -48,7 +48,7 @@ class GameTest < ActiveSupport::TestCase
   test "build_settlement adjacent to tile picks it up and decrements qty" do
     game = game_with_tile_at_2_7(qty: 2)
 
-    game.build_settlement(1, 7)
+    engine(game).build_settlement(1, 7)
     game.reload
 
     assert_equal 1, game.board_contents.tile_qty(2, 7)
@@ -61,7 +61,7 @@ class GameTest < ActiveSupport::TestCase
   test "taking the last tile keeps the board_contents entry at qty zero" do
     game = game_with_tile_at_2_7(qty: 1)
 
-    game.build_settlement(1, 7)
+    engine(game).build_settlement(1, 7)
     game.reload
 
     assert_not game.board_contents.empty?(2, 7), "entry must remain so the tile class is not lost"
@@ -75,7 +75,7 @@ class GameTest < ActiveSupport::TestCase
     chris.save
     game.reload  # clear association cache
 
-    game.build_settlement(1, 7)
+    engine(game).build_settlement(1, 7)
     game.reload
 
     assert_equal 2, game.board_contents.tile_qty(2, 7), "tile qty must be unchanged"
@@ -85,7 +85,7 @@ class GameTest < ActiveSupport::TestCase
   test "build_settlement does not pick up a tile whose qty is already zero" do
     game = game_with_tile_at_2_7(qty: 0)
 
-    game.build_settlement(1, 7)
+    engine(game).build_settlement(1, 7)
     game.reload
 
     assert_equal 0, game.board_contents.tile_qty(2, 7), "tile qty must stay at zero"
@@ -95,10 +95,10 @@ class GameTest < ActiveSupport::TestCase
 
   test "undo_last_move after a tile pickup restores tile and removes it from the player" do
     game = game_with_tile_at_2_7(qty: 2)
-    game.build_settlement(1, 7)
+    engine(game).build_settlement(1, 7)
     game.reload
 
-    game.undo_last_move
+    engine(game).undo_last_move
     game.reload
 
     assert_equal 2, game.board_contents.tile_qty(2, 7), "tile qty must be restored"
@@ -110,10 +110,10 @@ class GameTest < ActiveSupport::TestCase
 
   test "undo_last_move increments a zero-qty tile back to one" do
     game = game_with_tile_at_2_7(qty: 1)
-    game.build_settlement(1, 7)
+    engine(game).build_settlement(1, 7)
     game.reload
 
-    game.undo_last_move
+    engine(game).undo_last_move
     game.reload
 
     assert_equal 1, game.board_contents.tile_qty(2, 7)
@@ -131,7 +131,7 @@ class GameTest < ActiveSupport::TestCase
     chris.tiles = [ { "klass" => "PaddockTile", "from" => "[2, 7]", "used" => false } ]
     chris.save
 
-    game.select_action("paddock")
+    engine(game).select_action("paddock")
     game.reload
 
     assert_equal "paddock", game.current_action["type"]
@@ -144,7 +144,7 @@ class GameTest < ActiveSupport::TestCase
     game.current_action = { "type" => "paddock" }
     game.save
 
-    game.select_settlement(5, 5)
+    engine(game).select_settlement(5, 5)
     game.reload
 
     assert_equal "paddock", game.current_action["type"]
@@ -158,7 +158,7 @@ class GameTest < ActiveSupport::TestCase
     game.current_action = { "type" => "paddock", "from" => "[5, 5]" }
     game.save
 
-    game.move_settlement(5, 7)
+    engine(game).move_settlement(5, 7)
     game.reload
 
     assert game.board_contents.empty?(5, 5), "settlement must leave its old location"
@@ -177,7 +177,7 @@ class GameTest < ActiveSupport::TestCase
     chris.save
 
     # Move to [1,5] — a valid paddock hop, not adjacent to [2,7]
-    game.move_settlement(1, 5)
+    engine(game).move_settlement(1, 5)
 
     assert_empty game_players(:chris).reload.tiles
   end
@@ -196,7 +196,7 @@ class GameTest < ActiveSupport::TestCase
     game.current_action = { "type" => "paddock", "from" => "[1, 5]" }
     game.save
 
-    game.move_settlement(1, 7)
+    engine(game).move_settlement(1, 7)
     game.reload
 
     assert_equal 1, game.board_contents.tile_qty(2, 7)
@@ -215,7 +215,7 @@ class GameTest < ActiveSupport::TestCase
     chris.tiles = [ { "klass" => "OasisTile", "from" => "[2, 7]" } ]
     chris.save
 
-    game.move_settlement(1, 7)
+    engine(game).move_settlement(1, 7)
     game.reload
 
     assert_equal 2, game.board_contents.tile_qty(2, 7), "tile qty must be unchanged"
@@ -231,7 +231,7 @@ class GameTest < ActiveSupport::TestCase
     game.current_action = { "type" => "paddock", "from" => "[1, 5]" }
     game.save
 
-    game.move_settlement(1, 7)
+    engine(game).move_settlement(1, 7)
     game.reload
 
     assert_equal 0, game.board_contents.tile_qty(2, 7), "tile qty must stay at zero"
@@ -245,9 +245,9 @@ class GameTest < ActiveSupport::TestCase
     game.current_action = { "type" => "paddock", "from" => "[5, 5]" }
     game.save
 
-    game.move_settlement(5, 7)
+    engine(game).move_settlement(5, 7)
     game.reload
-    game.undo_last_move
+    engine(game).undo_last_move
     game.reload
 
     assert game.board_contents.empty?(5, 7), "settlement must leave the destination"
@@ -259,10 +259,10 @@ class GameTest < ActiveSupport::TestCase
 
   test "undo_last_move after select_action resets current_action to mandatory" do
     game = games(:game2player)
-    game.select_action("paddock")
+    engine(game).select_action("paddock")
     game.reload
 
-    game.undo_last_move
+    engine(game).undo_last_move
     game.reload
 
     assert_equal({ "type" => "mandatory" }, game.current_action)
@@ -276,9 +276,9 @@ class GameTest < ActiveSupport::TestCase
     game.current_action = { "type" => "paddock" }
     game.save
 
-    game.select_settlement(5, 5)
+    engine(game).select_settlement(5, 5)
     game.reload
-    game.undo_last_move
+    engine(game).undo_last_move
     game.reload
 
     assert_equal "paddock", game.current_action["type"]
@@ -291,7 +291,7 @@ class GameTest < ActiveSupport::TestCase
     game.current_action = { "type" => "paddock", "from" => "[5, 5]" }
     game.save
 
-    game.end_turn
+    engine(game).end_turn
     game.reload
 
     assert_equal({ "type" => "mandatory" }, game.current_action)
@@ -308,28 +308,28 @@ class GameTest < ActiveSupport::TestCase
   test "tile_activatable? is false when tile is used" do
     game = games(:game2player)
     tile = { "klass" => "PaddockTile", "from" => "[2, 18]", "used" => true }
-    assert_not game.tile_activatable?(tile)
+    assert_not engine(game).tile_activatable?(tile)
   end
 
   test "tile_activatable? is true when tile is unused and mandatory_count equals MANDATORY_COUNT" do
     game = games(:game2player)
     # mandatory_count starts at 3 = MANDATORY_COUNT in fixture
     tile = { "klass" => "PaddockTile", "from" => "[2, 18]", "used" => false }
-    assert game.tile_activatable?(tile)
+    assert engine(game).tile_activatable?(tile)
   end
 
   test "tile_activatable? is false when mandatory_count is mid-build" do
     game = games(:game2player)
     game.mandatory_count = 1
     tile = { "klass" => "PaddockTile", "from" => "[2, 18]", "used" => false }
-    assert_not game.tile_activatable?(tile)
+    assert_not engine(game).tile_activatable?(tile)
   end
 
   test "tile_activatable? is true when mandatory_count is 0" do
     game = games(:game2player)
     game.mandatory_count = 0
     tile = { "klass" => "PaddockTile", "from" => "[2, 18]", "used" => false }
-    assert game.tile_activatable?(tile)
+    assert engine(game).tile_activatable?(tile)
   end
 
   test "tile_activatable? is true when supply is 0 regardless of mandatory_count" do
@@ -337,7 +337,7 @@ class GameTest < ActiveSupport::TestCase
     game.mandatory_count = 1
     game.current_player.supply["settlements"] = 0
     tile = { "klass" => "PaddockTile", "from" => "[2, 18]", "used" => false }
-    assert game.tile_activatable?(tile)
+    assert engine(game).tile_activatable?(tile)
   end
 
   test "apply_tile_forfeit creates a forfeit_tile Move record for each forfeited tile" do
@@ -354,7 +354,7 @@ class GameTest < ActiveSupport::TestCase
     chris.tiles = [ { "klass" => "OasisTile", "from" => "[2, 7]", "used" => false } ]
     chris.save
 
-    game.move_settlement(1, 5)
+    engine(game).move_settlement(1, 5)
 
     forfeit_move = game.moves.find_by(action: "forfeit_tile")
     assert forfeit_move, "forfeit_tile Move must be created"
@@ -378,9 +378,9 @@ class GameTest < ActiveSupport::TestCase
     chris.tiles = [ { "klass" => "OasisTile", "from" => "[2, 7]", "used" => false } ]
     chris.save
 
-    game.move_settlement(1, 5)
+    engine(game).move_settlement(1, 5)
     game.reload
-    game.undo_last_move
+    engine(game).undo_last_move
     game.reload
 
     chris.reload
@@ -403,7 +403,7 @@ class GameTest < ActiveSupport::TestCase
     ]
     chris.save
 
-    game.move_settlement(5, 7)
+    engine(game).move_settlement(5, 7)
     chris.reload
 
     paddock_tiles = chris.tiles.select { |t| t["klass"] == "PaddockTile" }
@@ -425,9 +425,9 @@ class GameTest < ActiveSupport::TestCase
     ]
     chris.save
 
-    game.move_settlement(5, 7)
+    engine(game).move_settlement(5, 7)
     game.reload
-    game.undo_last_move
+    engine(game).undo_last_move
     chris.reload
 
     paddock_tile = chris.tiles.find { |t| t["klass"] == "PaddockTile" }
@@ -443,7 +443,7 @@ class GameTest < ActiveSupport::TestCase
     ]
     paula.save
 
-    game.end_turn
+    engine(game).end_turn
     paula.reload
 
     assert paula.tiles.all? { |t| t["used"] == false },
@@ -454,21 +454,21 @@ class GameTest < ActiveSupport::TestCase
     game = games(:game2player)
     game.mandatory_count = 0
     game.current_action = { "type" => "paddock" }
-    assert_not game.turn_endable?
+    assert_not engine(game).turn_endable?
   end
 
   test "turn_endable? returns false when paddock action has from selected" do
     game = games(:game2player)
     game.mandatory_count = 0
     game.current_action = { "type" => "paddock", "from" => "[5, 5]" }
-    assert_not game.turn_endable?
+    assert_not engine(game).turn_endable?
   end
 
   test "turn_endable? returns true when mandatory action is complete" do
     game = games(:game2player)
     game.mandatory_count = 0
     game.current_action = { "type" => "mandatory" }
-    assert game.turn_endable?
+    assert engine(game).turn_endable?
   end
 
   test "turn_endable? returns true when supply is 0 and action is mandatory" do
@@ -477,7 +477,7 @@ class GameTest < ActiveSupport::TestCase
     chris = game_players(:chris)
     chris.supply = { "settlements" => 0 }
     chris.save
-    assert game.turn_endable?
+    assert engine(game).turn_endable?
   end
 
   test "turn_state says 'must end their turn or select a tile' when mandatory done and tile is activatable" do
@@ -489,14 +489,14 @@ class GameTest < ActiveSupport::TestCase
     chris.save
     game.reload
 
-    assert_match(/must end their turn or select a tile/, game.turn_state)
+    assert_match(/must end their turn or select a tile/, engine(game).turn_state)
   end
 
   test "turn_state says 'must end their turn' without tile option when no activatable tiles" do
     game = games(:game2player)
 
-    assert_match(/must end their turn/, game.turn_state)
-    assert_no_match(/or select a tile/, game.turn_state)
+    assert_match(/must end their turn/, engine(game).turn_state)
+    assert_no_match(/or select a tile/, engine(game).turn_state)
   end
 
   test "turn_state includes 'or select a tile' at start of turn when player has an activatable tile" do
@@ -509,19 +509,19 @@ class GameTest < ActiveSupport::TestCase
     chris.save
     game.reload
 
-    assert_match(/or select a tile/, game.turn_state)
+    assert_match(/or select a tile/, engine(game).turn_state)
   end
 
   test "turn_state returns must move a settlement when paddock has no from" do
     game = games(:game2player)
     game.current_action = { "type" => "paddock" }
-    assert_match(/must move a settlement/, game.turn_state)
+    assert_match(/must move a settlement/, engine(game).turn_state)
   end
 
   test "turn_state returns must move a settlement when paddock has from set" do
     game = games(:game2player)
     game.current_action = { "type" => "paddock", "from" => "[5, 5]" }
-    assert_match(/must move a settlement/, game.turn_state)
+    assert_match(/must move a settlement/, engine(game).turn_state)
   end
 
   # Oasis tile action tests
@@ -537,7 +537,7 @@ class GameTest < ActiveSupport::TestCase
     chris.tiles = [ { "klass" => "OasisTile", "from" => "[2, 7]", "used" => false } ]
     chris.save
 
-    game.select_action("oasis")
+    engine(game).select_action("oasis")
     game.reload
 
     assert_equal "oasis", game.current_action["type"]
@@ -547,7 +547,7 @@ class GameTest < ActiveSupport::TestCase
     game = game_in_oasis_action
     chris = game_players(:chris)
 
-    game.activate_tile_build(0, 1)
+    engine(game).activate_tile_build(0, 1)
     game.reload
 
     assert_equal chris.order, game.board_contents.player_at(0, 1)
@@ -556,7 +556,7 @@ class GameTest < ActiveSupport::TestCase
   test "build_on_terrain decrements the player supply by one" do
     game = game_in_oasis_action
 
-    game.activate_tile_build(0, 1)
+    engine(game).activate_tile_build(0, 1)
 
     assert_equal 39, game_players(:chris).reload.supply["settlements"]
   end
@@ -564,7 +564,7 @@ class GameTest < ActiveSupport::TestCase
   test "build_on_terrain marks the activating tile as used" do
     game = game_in_oasis_action
 
-    game.activate_tile_build(0, 1)
+    engine(game).activate_tile_build(0, 1)
 
     oasis_tile = game_players(:chris).reload.tiles.find { |t| t["klass"] == "OasisTile" }
     assert oasis_tile["used"]
@@ -573,7 +573,7 @@ class GameTest < ActiveSupport::TestCase
   test "build_on_terrain resets current_action to mandatory" do
     game = game_in_oasis_action
 
-    game.activate_tile_build(0, 1)
+    engine(game).activate_tile_build(0, 1)
     game.reload
 
     assert_equal({ "type" => "mandatory" }, game.current_action)
@@ -595,7 +595,7 @@ class GameTest < ActiveSupport::TestCase
     chris.tiles = [ { "klass" => "OasisTile", "from" => "[2, 7]", "used" => false } ]
     chris.save
 
-    game.activate_tile_build(7, 6)
+    engine(game).activate_tile_build(7, 6)
     game.reload
 
     assert game.moves.exists?(action: "pick_up_tile"), "tile pickup must be triggered"
@@ -605,9 +605,9 @@ class GameTest < ActiveSupport::TestCase
   test "undo_last_move after build_on_terrain removes the settlement and restores supply" do
     game = game_in_oasis_action
 
-    game.activate_tile_build(0, 1)
+    engine(game).activate_tile_build(0, 1)
     game.reload
-    game.undo_last_move
+    engine(game).undo_last_move
     game.reload
 
     assert game.board_contents.empty?(0, 1)
@@ -617,9 +617,9 @@ class GameTest < ActiveSupport::TestCase
   test "undo_last_move after build_on_terrain unmarks the activating tile" do
     game = game_in_oasis_action
 
-    game.activate_tile_build(0, 1)
+    engine(game).activate_tile_build(0, 1)
     game.reload
-    game.undo_last_move
+    engine(game).undo_last_move
     game.reload
 
     oasis_tile = game_players(:chris).reload.tiles.find { |t| t["klass"] == "OasisTile" }
@@ -629,9 +629,9 @@ class GameTest < ActiveSupport::TestCase
   test "undo_last_move after build_on_terrain restores current_action to the tile action type" do
     game = game_in_oasis_action
 
-    game.activate_tile_build(0, 1)
+    engine(game).activate_tile_build(0, 1)
     game.reload
-    game.undo_last_move
+    engine(game).undo_last_move
     game.reload
 
     assert_equal "oasis", game.current_action["type"]
@@ -641,14 +641,14 @@ class GameTest < ActiveSupport::TestCase
   test "turn_state returns must build on a Desert space when oasis action" do
     game = games(:game2player)
     game.current_action = { "type" => "oasis" }
-    assert_match(/must build on a Desert space/, game.turn_state)
+    assert_match(/must build on a Desert space/, engine(game).turn_state)
   end
 
   test "turn_endable? returns false when oasis action is in progress" do
     game = games(:game2player)
     game.mandatory_count = 0
     game.current_action = { "type" => "oasis" }
-    assert_not game.turn_endable?
+    assert_not engine(game).turn_endable?
   end
 
   test "tile_activatable? returns true for unused OasisTile when Desert hexes exist" do
@@ -657,7 +657,7 @@ class GameTest < ActiveSupport::TestCase
     game.board_contents = BoardState.new
     game.save
     tile = { "klass" => "OasisTile", "from" => "[2, 7]", "used" => false }
-    assert game.tile_activatable?(tile)
+    assert engine(game).tile_activatable?(tile)
   end
 
   test "tile_activatable? returns false for unused OasisTile when all Desert hexes are occupied" do
@@ -676,7 +676,7 @@ class GameTest < ActiveSupport::TestCase
     end
     game.save
     tile = { "klass" => "OasisTile", "from" => "[2, 7]", "used" => false }
-    assert_not game.tile_activatable?(tile)
+    assert_not engine(game).tile_activatable?(tile)
   end
 
   # Move payload tests
@@ -687,7 +687,7 @@ class GameTest < ActiveSupport::TestCase
   test "build stores the terrain card played in payload" do
     game = game_with_tile_at_2_7(qty: 0)  # Oasis board, Chris hand "T", build at (1,7)
 
-    game.build_settlement(1, 7)
+    engine(game).build_settlement(1, 7)
 
     build_move = game.moves.find_by(action: "build")
     assert_equal "T", build_move.payload["card"]
@@ -697,7 +697,7 @@ class GameTest < ActiveSupport::TestCase
   test "build_on_terrain stores terrain card and tile_klass in payload" do
     game = game_in_oasis_action
 
-    game.activate_tile_build(0, 1)
+    engine(game).activate_tile_build(0, 1)
 
     build_move = game.moves.find_by(action: "build")
     assert_equal "D", build_move.payload["card"]
@@ -713,7 +713,7 @@ class GameTest < ActiveSupport::TestCase
     game.discard = []
     game.save
 
-    game.end_turn
+    engine(game).end_turn
 
     move = game.moves.find_by(action: "end_turn")
     assert_equal "G", move.payload["card_discarded"]
@@ -730,7 +730,7 @@ class GameTest < ActiveSupport::TestCase
     game.discard = [ "D", "F", "T" ]
     game.save
 
-    game.end_turn
+    engine(game).end_turn
 
     move = game.moves.find_by(action: "end_turn")
     assert_equal "G", move.payload["card_discarded"]
@@ -742,7 +742,7 @@ class GameTest < ActiveSupport::TestCase
   test "pick_up_tile stores tile klass and qty_before in payload" do
     game = game_with_tile_at_2_7(qty: 2)
 
-    game.build_settlement(1, 7)
+    engine(game).build_settlement(1, 7)
 
     pickup_move = game.moves.find_by(action: "pick_up_tile")
     assert_equal "OasisTile", pickup_move.payload["klass"]
@@ -752,7 +752,7 @@ class GameTest < ActiveSupport::TestCase
   test "pick_up_tile message uses correct article for vowel-initial tile names" do
     game = game_with_tile_at_2_7(qty: 2)
 
-    game.build_settlement(1, 7)
+    engine(game).build_settlement(1, 7)
 
     pickup_move = game.moves.find_by(action: "pick_up_tile")
     assert_match(/picked up an Oasis tile/, pickup_move.message)
@@ -771,7 +771,7 @@ class GameTest < ActiveSupport::TestCase
     chris.tiles = [ { "klass" => "OasisTile", "from" => "[2, 7]", "used" => false } ]
     chris.save
 
-    game.move_settlement(1, 5)
+    engine(game).move_settlement(1, 5)
 
     forfeit_move = game.moves.find_by(action: "forfeit_tile")
     assert_equal "OasisTile", forfeit_move.payload["klass"]
@@ -790,7 +790,7 @@ class GameTest < ActiveSupport::TestCase
     chris.tiles = [ { "klass" => "OasisTile", "from" => "[2, 7]", "used" => false } ]
     chris.save
 
-    game.move_settlement(1, 5)
+    engine(game).move_settlement(1, 5)
 
     forfeit_move = game.moves.find_by(action: "forfeit_tile")
     assert_match(/forfeited an oasis tile/, forfeit_move.message)
@@ -811,7 +811,7 @@ class GameTest < ActiveSupport::TestCase
     chris.tiles = [ { "klass" => "OasisTile", "from" => "[2, 7]", "used" => false } ]
     chris.save
 
-    game.move_settlement(1, 5)
+    engine(game).move_settlement(1, 5)
     game.reload
 
     # Remove the tile entry from board_contents so undo cannot read klass from it
@@ -820,7 +820,7 @@ class GameTest < ActiveSupport::TestCase
     game.board_contents = bc
     game.save
 
-    game.undo_last_move
+    engine(game).undo_last_move
     game.reload
 
     restored = game_players(:chris).reload.tiles
@@ -828,6 +828,10 @@ class GameTest < ActiveSupport::TestCase
   end
 
   private
+
+  def engine(game)
+    TurnEngine.new(game)
+  end
 
   # Returns a saved, in-progress game using the Oasis board with a single tile
   # entry at overall coordinate [2, 7] (the first Oasis location hex).
